STY: move EIA detection
Move the EIA detection functions into a new submodule.
